Atualiza raspador de Campos dos Goytacazes-RJ
Corrige nome do arquivo (estava sem stopword)

Atualiza raspador para adotar utils implementadas e faz pequenas correções
